Добавлена трех этапная аутентификация для сдачи секретных документов. Вам нужно: Быть офицером интекью, с определенной ДНК, и ввести пароль который бует отправлен вам как только вы проснетесь как Vanguard. Раставлены вендорматы на аванпостах в карго интекью

+++ mod_celadon/inteq_vendor/code/vendor.dm
@@ -4,6 +4,10 @@
 	icon = 'mod_celadon/_storage_icons/icons/items/misc/gun_vaucher.dmi'
 	icon_state = "inteq-voucher"
 
+GLOBAL_LIST_EMPTY(inteq_authorized_officers)
+GLOBAL_LIST_EMPTY(inteq_authorized_dna)
+GLOBAL_VAR_INIT(inteq_document_password, "")
+
 /obj/machinery/vending/security/marine/inteq
 	name = "\improper Inteq equipment vendor"
 	desc = "An Inteq Risk Management Group equipment vendor."
@@ -28,16 +32,50 @@
 	var/document_value_min = 6000
 	var/document_value_max = 12000
 	var/document_value = 0
+	resistance_flags = INDESTRUCTIBLE | LAVA_PROOF | FIRE_PROOF | UNACIDABLE | ACID_PROOF
 
 /obj/machinery/vending/security/marine/inteq/Initialize()
 	. = ..()
 	document_value = rand(document_value_min, document_value_max)
 
+/obj/machinery/vending/security/marine/inteq/ui_interact(mob/user, datum/tgui/ui)
+	return
+
+/obj/machinery/vending/security/marine/inteq/emag_act(mob/user)
+	to_chat(user, span_warning("[src] has advanced anti-tampering protection!"))
+	return FALSE
+
+/obj/machinery/vending/security/marine/inteq/screwdriver_act(mob/living/user, obj/item/I)
+	to_chat(user, span_warning("[src] is sealed and cannot be disassembled!"))
+	return TRUE
+
+/obj/machinery/vending/security/marine/inteq/wrench_act(mob/living/user, obj/item/I)
+	to_chat(user, span_warning("[src] is bolted to the floor and cannot be moved!"))
+	return TRUE
+
+/obj/machinery/vending/security/marine/inteq/attack_hand(mob/user)
+	to_chat(user, span_notice("[src] only accepts documents and vouchers. Insert them directly."))
+	return
+
 /obj/machinery/vending/security/marine/inteq/attackby(obj/item/I, mob/user, params)
 	if(istype(I, /obj/item/documents))
+		if(!(user.real_name in GLOB.inteq_authorized_officers))
+			to_chat(user, span_warning("Access denied. Only authorized Inteq officers can submit documents."))
+			return
+		if(ishuman(user))
+			var/mob/living/carbon/human/H = user
+			if(!H.dna || !(H.dna.unique_enzymes in GLOB.inteq_authorized_dna))
+				to_chat(user, span_warning("DNA verification failed. Access denied."))
+				log_admin("INTEQ VENDOR: [key_name(user)] failed DNA verification at [AREACOORD(src)].")
+				return
+		var/input_password = stripped_input(user, "Enter document submission password:", "Password Required", "")
+		if(!input_password || input_password != GLOB.inteq_document_password)
+			to_chat(user, span_warning("Incorrect password. Access denied."))
+			return
 		to_chat(user, span_notice("You insert [I] into [src]. The machine dispenses [document_value] credits."))
 		var/obj/item/holochip/holochip = new(drop_location(), document_value)
 		user.put_in_hands(holochip)
+		log_admin("INTEQ VENDOR: [key_name(user)] submitted documents using password '[input_password]' and received [document_value] credits at [AREACOORD(src)].")
 		qdel(I)
 		return
 	return ..()
